feat: add vue-template-lexer #617
