+++ src/index.js
@@ -9,3 +9,4 @@ export { default as HandlebarsLexer } from './lexers/handlebars-lexer.js'
 export { default as HTMLLexer } from './lexers/html-lexer.js'
 export { default as JavascriptLexer } from './lexers/javascript-lexer.js'
 export { default as JsxLexer } from './lexers/jsx-lexer.js'
+export { default as VueTemplateLexer } from './lexers/vue-template-lexer.js'
