+++ src/lexers/html-lexer.js
@@ -1,5 +1,4 @@
 import BaseLexer from './base-lexer.js'
-import JavascriptLexer from './javascript-lexer.js'
 import * as cheerio from 'cheerio'
 
 export default class HTMLLexer extends BaseLexer {
@@ -8,7 +7,6 @@ export default class HTMLLexer extends BaseLexer {
 
     this.attr = options.attr || 'data-i18n'
     this.optionAttr = options.optionAttr || 'data-i18n-options'
-    this.vueBindAttr = options.vueBindAttr || false
   }
 
   extract(content) {
@@ -40,26 +38,6 @@ export default class HTMLLexer extends BaseLexer {
         }
       }
     })
-
-    // support vue bind attribute
-    if (this.vueBindAttr) {
-      let keys = []
-      const jsLexer = new JavascriptLexer({ functions: this.functions })
-      // traverse all tags to filter bind attribute
-      $('*').each((index, node) => {
-        const attributes = node.attributes.filter(
-          (item) => item.name.startsWith(':') || item.name.startsWith('v-bind:')
-        )
-        if (attributes.length > 0) {
-          // there are calculation attributes.
-          attributes.forEach((content) => {
-            const items = jsLexer.extract(content.value)
-            keys.push(...items)
-          })
-        }
-      })
-      this.keys.push(...keys)
-    }
 
     return this.keys
   }
