+++ test/lexers/html-lexer.test.js
@@ -98,26 +98,4 @@ describe('HTMLLexer', () => {
     ])
     done()
   })
-
-  it('supports vue bind attributes', (done) => {
-    const Lexer = new HTMLLexer({
-      functions: ['t', '$t'],
-      vueBindAttr: true,
-    })
-    const content = `
-    <template>
-      <button :aria-label="t('b_a_l', 'button aria label')">
-        button label
-      </button>
-      <button v-bind:aria-label="$t('button v-bind aria label')">
-        button label form v-bind
-      </button>
-    </template>
-    `
-    assert.deepEqual(Lexer.extract(content), [
-      { key: 'b_a_l', defaultValue: 'button aria label' },
-      { key: 'button v-bind aria label' },
-    ])
-    done()
-  })
 })
